backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# AI & ML
openai==1.3.9
anthropic==0.7.8
sentence-transformers==2.2.2
numpy==1.26.2
scikit-learn==1.3.2

# Document Processing
PyPDF2==3.0.1
python-docx==0.8.11
pillow==10.1.0
pytesseract==0.3.10
python-magic==0.4.27

# Vector Database
pinecone-client==2.2.2

# Caching & Async
redis==5.0.1
aioredis==2.0.1
httpx==0.25.2

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.12.1
python-dotenv==1.0.0

# Monitoring & Logging
python-json-logger==2.0.7
prometheus-client==0.19.0
sentry-sdk==1.39.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0

# Development
black==23.12.0
flake8==6.1.0
pylint==3.0.3
mypy==1.7.1
isort==5.13.2
